rename tsx to ts in services files
