+++ src/services/sendAnswerMock.ts
@@ -1,4 +1,4 @@
-import {TeamType} from "./getTeamTokenMock.tsx";
+import {TeamType} from "./getTeamTokenMock.ts";
 
 export const sendAnswer = (team: TeamType, answer: string): void => {
     console.log("send answer", team.leader, answer);
